+++ tests/Api/ServiceTest.php
@@ -541,9 +541,6 @@ class ServiceTest extends BaseTestCase
         $result = $service->refund('dummy', 10);
         $this->assertInstanceOf(Entities\Transaction::class, $result);
 
-        $result = $service->cancel('dummy');
-        $this->assertInstanceOf(Entities\Transaction::class, $result);
-
         $result = $service->patch('dummy', []);
         $this->assertInstanceOf(Entities\Transaction::class, $result);
     }