+++ src/Entities/Invoice.php
@@ -402,4 +402,20 @@ final class Invoice extends Entity
     {
         return $this->setAttribute('notes', $value);
     }
+
+    /**
+     * @return string
+     */
+    public function getCreatedTime()
+    {
+        return $this->getAttribute('createdTime');
+    }
+
+    /**
+     * @return string
+     */
+    public function getRevision()
+    {
+        return $this->getAttribute('revision');
+    }
 }