+++ tests/Api/ServiceTest.php
@@ -541,9 +541,6 @@ class ServiceTest extends BaseTestCase
         $result = $service->refund('dummy', 10);
         $this->assertInstanceOf(Entities\Transaction::class, $result);
 
-        $result = $service->cancel('dummy');
-        $this->assertInstanceOf(Entities\Transaction::class, $result);
-
         $result = $service->patch('dummy', []);
         $this->assertInstanceOf(Entities\Transaction::class, $result);
     }
@@ -650,9 +647,6 @@ class ServiceTest extends BaseTestCase
         $result = $service->createFromToken(['token' => 'dummy'], $card);
         $this->assertInstanceOf(Entities\PaymentCard::class, $result);
 
-        $result = $service->authorize([], 'dummy');
-        $this->assertInstanceOf(Entities\PaymentCard::class, $result);
-
         $result = $service->deactivate('dummy');
         $this->assertInstanceOf(Entities\PaymentCard::class, $result);
     }
@@ -689,53 +683,6 @@ class ServiceTest extends BaseTestCase
 
         $entity = $service->update($resourceType, $this->getFakeValue('id', $entityClass), []);
         $this->assertInstanceOf($entityClass, $entity);
-    }
-
-    /**
-     * @test
-     */
-    public function layoutService()
-    {
-        $client = new Client(['apiKey' => 'QWERTY']);
-
-        /** @var CurlHandler|MockObject $handler */
-        $handler = $this->createMock(CurlHandler::class);
-        $layout = new Entities\Layout();
-        $layout->addItem([
-            'planId' => self::uuid(),
-            'starred' => true,
-        ]);
-        $layout->setItems([
-            [
-                'planId' => self::uuid(),
-                'starred' => true,
-            ],
-            new Entities\LayoutItem([
-                'planId' => self::uuid(),
-            ]),
-        ]);
-
-        $handler
-            ->expects($this->any())
-            ->method('__invoke')
-            ->will($this->returnValue(
-                $client
-                    ->createResponse()
-                    ->withHeader('Location', 'layouts/dummy')
-                    ->withBody(Psr7\stream_for(json_encode($layout)))
-            ));
-
-        $client = new Client([
-            'apiKey' => 'QWERTY',
-            'httpHandler' => $handler,
-        ]);
-
-        $service = $client->layouts();
-
-        $result = $service->create($layout);
-        $this->assertInstanceOf(Entities\Layout::class, $result);
-        $this->assertCount(2, $result->getItems());
-        $this->assertInstanceOf(Entities\LayoutItem::class, $result->getItems()[0]);
     }
 
     /**
@@ -900,9 +847,6 @@ class ServiceTest extends BaseTestCase
         $service = $client->payPalAccounts();
 
         $result = $service->deactivate('dummy');
-        $this->assertInstanceOf(Entities\PayPalAccount::class, $result);
-
-        $result = $service->activate([], 'dummy');
         $this->assertInstanceOf(Entities\PayPalAccount::class, $result);
     }
 
@@ -1207,11 +1151,6 @@ class ServiceTest extends BaseTestCase
                 'invoices',
                 Services\InvoiceService::class,
                 Entities\Invoice::class,
-            ],
-            [
-                'layouts',
-                Services\LayoutService::class,
-                Entities\Layout::class,
             ],
             [
                 'paymentCards',