+++ CHANGELOG.md
@@ -35,6 +35,9 @@ Security - in case of vulnerabilities.
 
 - [x] Removed `put` method from `OrganizationService::create()` method.
 - [x] Removed `OrganizationService::delete()` method.
+- [x] Removed `PaymentCardService::authorize()` method.
+- [x] Removed `PayPalAccountService::activate()` method.
+- [x] Removed `PaymentCardAuthorization` entity.
 
 ## [2.11.0] 2021-03-03